debug booking

+++ src/js/components/BaseWidget.js
@@ -23,7 +23,7 @@ class BaseWidget {
     thisWidget.value = value;
   }
   parseValue(value) {
-    return parseInt(value);
+    return parseFloat(value);
   }
   isValid(value) {
     return !isNaN(value);
